Added delegated grant support for RecordsQery
