+++ json-schemas/authorization-delegated-grant.json
@@ -1,0 +1,14 @@
+{
+  "$id": "https://identity.foundation/dwn/json-schemas/authorization-delegated-grant.json",
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "type": "object",
+  "additionalProperties": false,
+  "properties": {
+    "signature": {
+      "$ref": "https://identity.foundation/dwn/json-schemas/general-jws.json"
+    },
+    "authorDelegatedGrant": {
+      "$ref": "https://identity.foundation/dwn/json-schemas/permissions-grant.json"
+    }
+  }
+}
