+++ src/interfaces/records-query.ts
@@ -1,7 +1,9 @@
-import type { Pagination } from '../types/message-types.js';
+import type { DelegatedGrantMessage } from '../types/permissions-types.js';
 import type { Signer } from '../types/signer.js';
+import type { GenericMessage, GenericSignaturePayload, Pagination } from '../types/message-types.js';
 import type { RecordsFilter, RecordsQueryDescriptor, RecordsQueryMessage } from '../types/records-types.js';
 
+import { Jws } from '../utils/jws.js';
 import { Message } from '../core/message.js';
 import { Records } from '../utils/records.js';
 import { removeUndefinedProperties } from '../utils/object.js';
@@ -25,9 +27,59 @@ export type RecordsQueryOptions = {
   pagination?: Pagination;
   signer?: Signer;
   protocolRole?: string;
+
+  /**
+   * The delegated grant to sign on behalf of the logical author, which is the grantor (`grantedBy`) of the delegated grant.
+   */
+  delegatedGrant?: DelegatedGrantMessage;
 };
 
-export class RecordsQuery extends Message<RecordsQueryMessage> {
+/**
+ * A class representing a RecordsQuery DWN message.
+ */
+export class RecordsQuery {
+  private _message: RecordsQueryMessage;
+  /**
+   * Valid JSON message representing this RecordsQuery.
+   */
+  public get message(): RecordsQueryMessage {
+    return this._message as RecordsQueryMessage;
+  }
+
+  private _author: string | undefined;
+  /**
+   * DID of the logical author of this message.
+   * NOTE: we say "logical" author because a message can be signed by a delegate of the actual author,
+   * in which case the author DID would not be the same as the signer/delegate DID,
+   * but be the DID of the grantor (`grantedBy`) of the delegated grant presented.
+   */
+  public get author(): string | undefined {
+    return this._author;
+  }
+
+  private _signaturePayload: GenericSignaturePayload | undefined;
+  /**
+   * Decoded payload of the signature of this message.
+   */
+  public get signaturePayload(): GenericSignaturePayload | undefined {
+    return this._signaturePayload;
+  }
+
+  private constructor(message: RecordsQueryMessage) {
+    this._message = message;
+
+    if (message.authorization !== undefined) {
+      // if the message authorization contains author delegated grant, the author would be the grantor of the grant
+      // else the author would be the signer of the message
+      if (message.authorization.authorDelegatedGrant !== undefined) {
+        this._author = Message.getSigner(message.authorization.authorDelegatedGrant);
+      } else {
+        this._author = Message.getSigner(message as GenericMessage);
+      }
+
+      this._signaturePayload = Jws.decodePlainObjectPayload(message.authorization.signature);
+    }
+  }
 
   public static async parse(message: RecordsQueryMessage): Promise<RecordsQuery> {
     let authorizationPayload;
@@ -72,7 +124,7 @@ export class RecordsQuery extends Message<RecordsQueryMessage> {
     const signer = options.signer;
     let authorization;
     if (signer) {
-      authorization = await Message.createAuthorization(descriptor, signer, { protocolRole: options.protocolRole });
+      authorization = await Message.createAuthorization(descriptor, signer, { protocolRole: options.protocolRole }, options.delegatedGrant);
     }
     const message = { descriptor, authorization };
