+++ tests/handlers/records-write.spec.ts
@@ -476,262 +476,6 @@ export function testRecordsWriteHandler(): void {
         });
       });
 
-      describe('delegated grant tests', () => {
-        it('should allow entity invoking a valid delegated grant to write', async () => {
-          // scenario: Alice creates a delegated grant for device X and device Y,
-          // device X and Y can both use their grants to write a message to Bob's DWN as Alice
-          // all party should treat messages written by device X and Y as if they were written by Alice
-          const alice = await DidKeyResolver.generate();
-          const deviceX = await DidKeyResolver.generate();
-          const deviceY = await DidKeyResolver.generate();
-          const bob = await DidKeyResolver.generate();
-
-          // Bob has the message protocol installed
-          const protocolDefinition = messageProtocolDefinition;
-          const protocol = protocolDefinition.protocol;
-          const protocolsConfig = await TestDataGenerator.generateProtocolsConfigure({
-            author: bob,
-            protocolDefinition
-          });
-          const protocolConfigureReply = await dwn.processMessage(bob.did, protocolsConfig.message);
-          expect(protocolConfigureReply.status.code).to.equal(202);
-
-          // Alice creates a delegated grant for device X and device Y
-          const scope: PermissionScope = {
-            interface : DwnInterfaceName.Records,
-            method    : DwnMethodName.Write,
-            protocol
-          };
-
-          const deviceXGrant = await PermissionsGrant.create({
-            delegated   : true, // this is a delegated grant
-            dateExpires : Time.createOffsetTimestamp({ seconds: 100 }),
-            description : 'Allow to write to message protocol',
-            grantedBy   : alice.did,
-            grantedTo   : deviceX.did,
-            grantedFor  : alice.did,
-            scope       : scope,
-            signer      : Jws.createSigner(alice)
-          });
-
-          const deviceYGrant = await PermissionsGrant.create({
-            delegated   : true, // this is a delegated grant
-            dateExpires : Time.createOffsetTimestamp({ seconds: 100 }),
-            description : 'Allow to write to message protocol',
-            grantedBy   : alice.did,
-            grantedTo   : deviceY.did,
-            grantedFor  : alice.did,
-            scope       : scope,
-            signer      : Jws.createSigner(alice)
-          });
-
-          // generate a `RecordsWrite` message from device X and write to Bob's DWN
-          const deviceXData = new TextEncoder().encode('message from device X');
-          const deviceXDataStream = DataStream.fromBytes(deviceXData);
-          const messageByDeviceX = await RecordsWrite.create({
-            signer         : Jws.createSigner(deviceX),
-            delegatedGrant : deviceXGrant.asDelegatedGrant(),
-            protocol,
-            protocolPath   : 'message', // this comes from `types` in protocol definition
-            schema         : protocolDefinition.types.message.schema,
-            dataFormat     : protocolDefinition.types.message.dataFormats[0],
-            data           : deviceXData
-          });
-
-          const deviceXWriteReply = await dwn.processMessage(bob.did, messageByDeviceX.message, deviceXDataStream);
-          expect(deviceXWriteReply.status.code).to.equal(202);
-
-          // verify the message by device X got written to Bob's DWN, AND Alice is the logical author
-          const recordsQueryByBob = await TestDataGenerator.generateRecordsQuery({
-            author : bob,
-            filter : { protocol }
-          });
-          const bobRecordsQueryReply = await dwn.processMessage(bob.did, recordsQueryByBob.message);
-          expect(bobRecordsQueryReply.status.code).to.equal(200);
-          expect(bobRecordsQueryReply.entries?.length).to.equal(1);
-
-          const fetchedDeviceXWriteEntry =bobRecordsQueryReply.entries![0];
-          expect(fetchedDeviceXWriteEntry.encodedData).to.equal(base64url.baseEncode(deviceXData));
-
-          const fetchedDeviceXWrite = await RecordsWrite.parse(fetchedDeviceXWriteEntry);
-          expect(fetchedDeviceXWrite.author).to.equal(alice.did);
-
-          // generate a new message by device Y updating the existing record device X created, and write to Bob's DWN
-          const deviceYData = new TextEncoder().encode('message from device Y');
-          const deviceYDataStream = DataStream.fromBytes(deviceYData);
-          const messageByDeviceY = await RecordsWrite.createFrom({
-            recordsWriteMessage : fetchedDeviceXWrite.message,
-            data                : deviceYData,
-            signer              : Jws.createSigner(deviceY),
-            delegatedGrant      : deviceYGrant.asDelegatedGrant(),
-          });
-
-          const deviceYWriteReply = await dwn.processMessage(bob.did, messageByDeviceY.message, deviceYDataStream);
-          expect(deviceYWriteReply.status.code).to.equal(202);
-
-          // verify the message by device Y got written to Bob's DWN, AND Alice is the logical author
-          const bobRecordsQueryReply2 = await dwn.processMessage(bob.did, recordsQueryByBob.message);
-          expect(bobRecordsQueryReply2.status.code).to.equal(200);
-          expect(bobRecordsQueryReply2.entries?.length).to.equal(1);
-
-          const fetchedDeviceYWriteEntry =bobRecordsQueryReply2.entries![0];
-          expect(fetchedDeviceYWriteEntry.encodedData).to.equal(base64url.baseEncode(deviceYData));
-
-          const fetchedDeviceYWrite = await RecordsWrite.parse(fetchedDeviceYWriteEntry);
-          expect(fetchedDeviceYWrite.author).to.equal(alice.did);
-        });
-
-        xit('should allow entity invoking a valid delegated grant to read', async () => {
-        });
-
-        xit('should allow entity invoking a valid delegated grant to query', async () => {
-        });
-
-        xit('should allow entity invoking a valid delegated grant to delete', async () => {
-        });
-
-        xit('should not allow entity using a non-delegated grant as a delegated grant to invoke write', async () => {
-        });
-
-        xit('should not allow entity using a non-delegated grant as a delegated grant to invoke read', async () => {
-        });
-
-        xit('should not allow entity using a non-delegated grant as a delegated grant to invoke query', async () => {
-        });
-
-        xit('should not allow entity using a non-delegated grant as a delegated grant to invoke delete', async () => {
-        });
-
-        it('should fail if invoking a delegated grant that is issued to a different entity to write', async () => {
-          // scenario:
-          // 1. Alice creates a delegated grant for Bob
-          // 2. Carol starts a chat thread with Alice
-          // 3. Daniel stole Bob's delegated grant and attempts to write to Carol's DWN
-          // 4. Daniel's attempt should fail
-          const alice = await DidKeyResolver.generate();
-          const bob = await DidKeyResolver.generate();
-          const carol = await DidKeyResolver.generate();
-          const daniel = await DidKeyResolver.generate();
-
-          // Carol has the chat protocol installed
-          const protocolDefinition = threadRoleProtocolDefinition;
-          const protocol = threadRoleProtocolDefinition.protocol;
-          const protocolsConfig = await TestDataGenerator.generateProtocolsConfigure({
-            author: carol,
-            protocolDefinition
-          });
-          const protocolsConfigureReply = await dwn.processMessage(carol.did, protocolsConfig.message);
-          expect(protocolsConfigureReply.status.code).to.equal(202);
-
-          // Carol starts a chat thread
-          const threadRecord = await TestDataGenerator.generateRecordsWrite({
-            author       : carol,
-            protocol     : protocolDefinition.protocol,
-            protocolPath : 'thread',
-          });
-          const threadRoleReply = await dwn.processMessage(carol.did, threadRecord.message, threadRecord.dataStream);
-          expect(threadRoleReply.status.code).to.equal(202);
-
-          // Carol adds Alice as a participant in the thread
-          const participantRoleRecord = await TestDataGenerator.generateRecordsWrite({
-            author       : carol,
-            recipient    : alice.did,
-            protocol     : protocolDefinition.protocol,
-            protocolPath : 'thread/participant',
-            contextId    : threadRecord.message.contextId,
-            parentId     : threadRecord.message.recordId,
-            data         : new TextEncoder().encode('Alice is my friend'),
-          });
-          const participantRoleReply = await dwn.processMessage(carol.did, participantRoleRecord.message, participantRoleRecord.dataStream);
-          expect(participantRoleReply.status.code).to.equal(202);
-
-          // Alice creates a delegated grant for Bob to act as Alice.
-          const scope: PermissionScope = {
-            interface : DwnInterfaceName.Records,
-            method    : DwnMethodName.Write,
-            protocol
-          };
-
-          const grantToBob = await PermissionsGrant.create({
-            delegated   : true, // this is a delegated grant
-            dateExpires : Time.createOffsetTimestamp({ seconds: 100 }),
-            description : 'Allow to Bob write as me in chat protocol',
-            grantedBy   : alice.did,
-            grantedTo   : bob.did,
-            grantedFor  : alice.did,
-            scope       : scope,
-            signer      : Jws.createSigner(alice)
-          });
-
-          // Sanity check that Bob can write a chat message as Alice by invoking the delegated grant
-          const messageByBobAsAlice = new TextEncoder().encode('Message from Bob as Alice');
-          const writeByBobAsAlice = await RecordsWrite.create({
-            signer         : Jws.createSigner(bob),
-            delegatedGrant : grantToBob.asDelegatedGrant(),
-            protocolRole   : 'thread/participant',
-            protocol,
-            protocolPath   : 'thread/chat', // this comes from `types` in protocol definition
-            schema         : 'unused',
-            dataFormat     : 'unused',
-            contextId      : threadRecord.message.contextId,
-            parentId       : threadRecord.message.recordId,
-            data           : messageByBobAsAlice
-          });
-
-          const bobWriteReply = await dwn.processMessage(carol.did, writeByBobAsAlice.message, DataStream.fromBytes(messageByBobAsAlice));
-          expect(bobWriteReply.status.code).to.equal(202);
-
-          // Verify that Daniel cannot write a chat message as Alice by invoking the delegated grant granted to Bob
-          const messageByDanielAsAlice = new TextEncoder().encode('Message from Daniel as Alice');
-          const writeByDanielAsAlice = await RecordsWrite.create({
-            signer         : Jws.createSigner(daniel),
-            delegatedGrant : grantToBob.asDelegatedGrant(),
-            protocolRole   : 'thread/participant',
-            protocol,
-            protocolPath   : 'thread/chat', // this comes from `types` in protocol definition
-            schema         : 'unused',
-            dataFormat     : 'unused',
-            contextId      : threadRecord.message.contextId,
-            parentId       : threadRecord.message.recordId,
-            data           : messageByDanielAsAlice
-          });
-
-          const danielWriteReply = await dwn.processMessage(carol.did, writeByDanielAsAlice.message, DataStream.fromBytes(messageByDanielAsAlice));
-          expect(danielWriteReply.status.code).to.equal(400);
-          expect(danielWriteReply.status.detail).to.contain(DwnErrorCode.RecordsWriteValidateIntegrityGrantedToAndSignerMismatch);
-        });
-
-        it('should fail if invoking a delegated grant that is issued to a different entity to read', async () => {
-        });
-
-        it('should fail if invoking a delegated grant that is issued to a different entity to query', async () => {
-        });
-
-        it('should fail if invoking a delegated grant that is issued to a different entity to delete', async () => {
-        });
-
-        xit('should evaluate scoping correctly when invoking a delegated grant to write', async () => {
-        });
-
-        xit('should evaluate scoping correctly when invoking a delegated grant to read', async () => {
-        });
-
-        xit('should evaluate scoping correctly when invoking a delegated grant to query', async () => {
-        });
-
-        xit('should evaluate scoping correctly when invoking a delegated grant to delete', async () => {
-        });
-
-        xit('should not be able to create a RecordsWrite with a non-delegated grant assigned to `authorDelegatedGrant`', async () => {
-        });
-
-        xit('should fail if presented with a delegated grant with invalid grantor signature', async () => {
-        });
-
-        xit('should fail if presented with a delegated grant with mismatching grant ID in the payload of the message signature', async () => {
-        });
-      });
-
       describe('should inherit data from previous RecordsWrite given a matching dataCid and dataSize and no dataStream', () => {
         it('with data above the threshold for encodedData', async () => {
           const { message, author, dataStream, dataBytes } = await TestDataGenerator.generateRecordsWrite({
