+++ tests/test-suite.ts
@@ -1,7 +1,8 @@
 import type { DataStore, EventLog, MessageStore } from '../src/index.js';
 
+import { testDelegatedGrantScenarios } from './scenarios/delegated-grant.spec.js';
 import { testDwnClass } from './dwn.spec.js';
-import { testEndToEndScenarios } from './end-to-end-tests.spec.js';
+import { testEndToEndScenarios } from './scenarios/end-to-end-tests.spec.js';
 import { testEventsGetHandler } from './handlers/events-get.spec.js';
 import { testMessagesGetHandler } from './handlers/messages-get.spec.js';
 import { testMessageStore } from './store/message-store.spec.js';
@@ -33,6 +34,7 @@ export class TestSuite {
     testDwnClass();
     testMessageStore();
 
+    // handler tests
     testEventsGetHandler();
     testMessagesGetHandler();
     testPermissionsGrantHandler();
@@ -44,6 +46,8 @@ export class TestSuite {
     testRecordsReadHandler();
     testRecordsWriteHandler();
 
+    // scenario tests
+    testDelegatedGrantScenarios();
     testEndToEndScenarios();
   }
 }
